Epclauncher added.

+++ release.bat
@@ -1,13 +1,14 @@
 set PYTHON=python
 
 
-rd /S /Q venv
+if exist venv rd /S /Q venv
+if exist dist rd /S /Q dist
+if exist build rd /S /Q build
+if exist release rd /S /Q releaseв
 %PYTHON% -m venv venv
 venv\Scripts\python -m pip install --upgrade pip
 venv\Scripts\python -m pip install -r requirements.txt
-
 venv\Scripts\python -m pip install pyinstaller
-rd /S /Q dist
 venv\Scripts\pyinstaller main.py ^
 --add-data "venv\Lib\site-packages\epcore\ivmeasurer\ivm-win64\ivm.dll;." ^
 --add-data "venv\Lib\site-packages\epcore\doc\p10_elements.schema.json;epcore\doc" ^
@@ -17,4 +18,13 @@ venv\Scripts\pyinstaller main.py ^
 --add-data "media\*;media" ^
 --add-data "gui\*;gui" ^
 --icon=media\ico.ico
-rd /S /Q build
+
+xcopy release_templates\* dist\* /S /E
+rename dist release
+cd release
+rename main eplab
+cd ..
+if exist build rd /S /Q build
+if exist dist rd /S /Q dist
+if exist venv rd /S /Q venv
+if exist *.spec del *.spec
